toy_project/common/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cw-common"
version = "0.1.0"
description = "Shared observability utilities for CW Monitoring"
requires-python = ">=3.11"
dependencies = [
    "opentelemetry-api>=1.25.0",
    "opentelemetry-sdk>=1.25.0",
    "opentelemetry-exporter-otlp-proto-http>=1.25.0",
    "opentelemetry-instrumentation-logging>=0.46b0",
    "prometheus-client>=0.21.0",
    "python-json-logger>=2.0.7",
    "starlette>=0.27.0",
]
